🎩🐯 Checkpoint ./acarkomutlar/acar-isim.js:23664531/1167 ./acarkomutlar/acar-jail.js:23664531/685 ./acarkomutlar/acar-kayıt.js:23664531/472 ./acarkomutlar/acar-kayıtsız.js:23664531/740 ./acarkomutlar/acar-küfürengel.js:23664531/427 ./acarkomutlar/acar-mute.js:23664531/248 ./acarkomutlar/acar-ping.js:23664531/1165 ./acarkomutlar/acar-reboot.js:23664531/639 ./acarkomutlar/acar-reklamban-aç.js:23664531/427 ./acarkomutlar/acar-say.js:23664531/430 ./acarkomutlar/acar-sesmute.js:23664531/215 ./acarkomutlar/acar-taşı.js:23664531/215 ./acarkomutlar/acar-temizle.js:23664531/427 ./acarkomutlar/acar-unban.js:23664531/215 ./acarkomutlar/acar-unjail.js:23664531/795 ./acarkomutlar/acar-unmute.js:23664531/247 ./acarkomutlar/acar-yetki.js:23664531/229 ./acarkomutlar/acar-ytaşı.js:23664531/215

+++ acarkomutlar/acar-reboot.js
@@ -5,6 +5,7 @@ let acar = require('../acar/botayarlari.json');
 
 
 module.exports.run = async (bot, message, args) => {
+  if(message.channel.id !== acarayarlar.botkomutkanalid ) return message.channel.send("Lütfen komutu kullanmak için <#"+ acarayarlar.botkomutkanalid + "> kanalını kullanınız!").then(message => message.delete(3000))
    message.delete()
     if(message.author.id !== acar.sahip)  return message.channel.send("Bu Komutu Sadece Sahibim Kullanabilir")
